Move types out of `src/components/ui`
