+++ frontend/src/components/ui/Button/Button.tsx
@@ -2,8 +2,7 @@ import * as React from "react";
 import { Link, LinkProps } from "react-router";
 
 import { cn } from "@/lib/util/classnames";
-
-import { ButtonVariant, Size } from "@kiesraad/ui";
+import { ButtonVariant, Size } from "@/types/ui";
 
 import cls from "./Button.module.css";
