+++ frontend/src/components/ui/Icon/Icon.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 
 import { cn } from "@/lib/util/classnames";
+import { Size } from "@/types/ui";
 
-import { ProgressBarColorClass, Size } from "@kiesraad/ui";
-
+import { ProgressBarColorClass } from "../ProgressBar";
 import cls from "./Icon.module.css";
 
 export interface IconProps {
