+++ frontend/src/components/ui/ProgressList/ProgressList.tsx
@@ -2,8 +2,9 @@ import * as React from "react";
 
 import { cn } from "@/lib/util/classnames";
 import { domtoren } from "@/lib/util/domtoren";
+import { MenuStatus } from "@/types/ui";
 
-import { MenuStatus, renderStatusIcon } from "@kiesraad/ui";
+import { renderStatusIcon } from "@kiesraad/ui";
 
 import cls from "./ProgressList.module.css";
 import { ProgressListScroll } from "./ProgressListScroll";
