+++ frontend/src/components/ui/index.ts
@@ -25,7 +25,6 @@ export * from "./ProgressList/ProgressList";
 export * from "./Spinner/Spinner";
 export * from "./StatusList/StatusList";
 export * from "./Table/Table";
-export * from "./ui.types.d";
 export * from "./util";
 export * from "./Toolbar/Toolbar";
 export * from "./Pagination/Pagination";
